The rule-based solver needs to keep track of much more data.
A few of the rules can only prohibit values from cells, so
we need to remember "value V forbidden in cell C" as well;
but it's easier to keep track of the still allowed values.

And while the rows, columns,  and quadrants are mostly static
[with the exception that instead of plain quadrants sometimes
arbitrary connected shapes are used],  recalculating them all
the time isn't that nice.